# api/openapi.yaml
openapi: 3.0.0
info:
  title: EKKE Token Webshop API
  version: 1.0.0
servers:
  - url: http://localhost:4201
    description: Local server
paths:
...
  /user/edit-user/{id}:
    post:
      summary: Edits a user by id.
      tags:
        - userApi
      operationId: editUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                name:
                  type: string
                neptun:
                  type: string
                email:
                  type: string
                password:
                  type: string
  /user/delete-user/{id}:
    delete:
      summary: Deletes a user by id.
      tags:
        - userApi
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
...
  /product/edit-product/{id}:
    post:
      summary: Edits a product by id.
      tags:
        - productApi
      operationId: editProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                variant_id:
                  type: string
                name:
                  type: string
                type:
                  type: string
                style:
                  type: string
                color:
                  type: string
                size:
                  type: string
                price:
                  type: number
                quantity:
                  type: number
  /product/delete-product/{id}:
    delete:
      summary: Deletes a product by id.
      tags:
        - productApi
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
...